Code
Task:
Complete the Difference class by writing the following:
A class constructor that takes an array of integers as a parameter and saves it to the  instance variable.
A computeDifference method that finds the maximum absolute difference between any  numbers in  and stores it in the  instance variable.
Input Format:
You are not responsible for reading any input from stdin. The locked Solution class in your editor reads in 2 lines of input; the first line contains N, and the second line describes the elements array.
Output Format:
You are not responsible for printing any output; the Solution class will print the value of the maximumDifference instance variable.