Code
Task:
The AdvancedArithmetic interface and the method declaration for the abstract divisorSum(n) method are provided for you in the editor below.
Complete the implementation of Calculator class, which implements the AdvancedArithmetic interface. The implementation for the divisorSum(n) method must return the sum of all divisors of n.
Input Format:
A single line containing an integer, n.
Output Format:
You are not responsible for printing anything to stdout. The locked template code in the editor below will call your code and print the necessary output.